mosu-46 refactor: coderabbit 리뷰 반영
