+++ src/main/java/life/mosu/mosuserver/application/application/ApplicationService.java
@@ -11,10 +11,10 @@ import life.mosu.mosuserver.domain.application.ApplicationSchoolJpaRepository;
 import life.mosu.mosuserver.domain.applicationschool.ApplicationSchoolJpaEntity;
 import life.mosu.mosuserver.global.exception.CustomRuntimeException;
 import life.mosu.mosuserver.global.exception.ErrorCode;
+import life.mosu.mosuserver.global.util.FileRequest;
 import life.mosu.mosuserver.presentation.application.dto.ApplicationRequest;
 import life.mosu.mosuserver.presentation.application.dto.ApplicationResponse;
 import life.mosu.mosuserver.presentation.application.dto.ApplicationSchoolRequest;
-import life.mosu.mosuserver.presentation.faq.dto.FileRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
