+++ src/main/java/life/mosu/mosuserver/application/inquiry/InquiryAnswerAttachmentService.java
@@ -3,13 +3,13 @@ package life.mosu.mosuserver.application.inquiry;
 import java.time.Duration;
 import java.util.List;
 import life.mosu.mosuserver.domain.inquiryAnswer.InquiryAnswerAttachmentEntity;
-import life.mosu.mosuserver.domain.inquiryAnswer.InquiryAnswerAttachmentReposiory;
+import life.mosu.mosuserver.domain.inquiryAnswer.InquiryAnswerAttachmentRepository;
 import life.mosu.mosuserver.domain.inquiryAnswer.InquiryAnswerJpaEntity;
+import life.mosu.mosuserver.global.util.FileRequest;
 import life.mosu.mosuserver.infra.property.S3Properties;
 import life.mosu.mosuserver.infra.storage.FileUploadHelper;
 import life.mosu.mosuserver.infra.storage.application.AttachmentService;
 import life.mosu.mosuserver.infra.storage.application.S3Service;
-import life.mosu.mosuserver.presentation.faq.dto.FileRequest;
 import life.mosu.mosuserver.presentation.inquiry.dto.InquiryDetailResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -20,33 +20,33 @@ public class InquiryAnswerAttachmentService implements
         AttachmentService<InquiryAnswerJpaEntity, FileRequest> {
 
     private final S3Properties s3Properties;
-    private final InquiryAnswerAttachmentReposiory attachmentReposiory;
+    private final InquiryAnswerAttachmentRepository attachmentRepository;
     private final FileUploadHelper fileUploadHelper;
     private final S3Service s3Service;
 
     @Override
-    public void createAttachment(List<FileRequest> requests, InquiryAnswerJpaEntity inquiryEntity) {
+    public void createAttachment(List<FileRequest> requests, InquiryAnswerJpaEntity answerEntity) {
         if (requests == null) {
             return;
         }
-        Long inquiryId = inquiryEntity.getId();
-        saveAttachments(requests, inquiryId);
+        Long answerId = answerEntity.getId();
+        saveAttachments(requests, answerId);
     }
 
 
     @Override
     public void deleteAttachment(InquiryAnswerJpaEntity entity) {
-        List<InquiryAnswerAttachmentEntity> attachments = attachmentReposiory.findAllByInquiryAnswerId(
+        List<InquiryAnswerAttachmentEntity> attachments = attachmentRepository.findAllByInquiryAnswerId(
                 entity.getId());
 
-        attachmentReposiory.deleteAll(attachments);
+        attachmentRepository.deleteAll(attachments);
     }
 
 
     public List<InquiryDetailResponse.AttachmentDetailResponse> toAttachmentResponses(
             InquiryAnswerJpaEntity inquiry) {
 
-        List<InquiryAnswerAttachmentEntity> attachments = attachmentReposiory.findAllByInquiryAnswerId(
+        List<InquiryAnswerAttachmentEntity> attachments = attachmentRepository.findAllByInquiryAnswerId(
                 inquiry.getId());
 
         return attachments.stream()
@@ -84,7 +84,7 @@ public class InquiryAnswerAttachmentService implements
     private void saveAttachments(List<FileRequest> requests, Long inquiryId) {
         requests.forEach(req -> {
             fileUploadHelper.updateTag(req.s3Key());
-            attachmentReposiory.save(req.toInquiryAnswerAttachmentEntity(
+            attachmentRepository.save(req.toInquiryAnswerAttachmentEntity(
                     req.fileName(), req.s3Key(), inquiryId
             ));
         });
