+++ src/main/java/life/mosu/mosuserver/application/inquiry/InquiryAttachmentService.java
@@ -5,11 +5,11 @@ import java.util.List;
 import life.mosu.mosuserver.domain.inquiry.InquiryAttachmentJpaEntity;
 import life.mosu.mosuserver.domain.inquiry.InquiryAttachmentRepository;
 import life.mosu.mosuserver.domain.inquiry.InquiryJpaEntity;
+import life.mosu.mosuserver.global.util.FileRequest;
 import life.mosu.mosuserver.infra.property.S3Properties;
 import life.mosu.mosuserver.infra.storage.FileUploadHelper;
 import life.mosu.mosuserver.infra.storage.application.AttachmentService;
 import life.mosu.mosuserver.infra.storage.application.S3Service;
-import life.mosu.mosuserver.presentation.faq.dto.FileRequest;
 import life.mosu.mosuserver.presentation.inquiry.dto.InquiryDetailResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
