+++ src/main/java/life/mosu/mosuserver/domain/inquiryAnswer/InquiryAnswerAttachmentRepository.java
@@ -3,7 +3,7 @@ package life.mosu.mosuserver.domain.inquiryAnswer;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InquiryAnswerAttachmentReposiory extends
+public interface InquiryAnswerAttachmentRepository extends
         JpaRepository<InquiryAnswerAttachmentEntity, Long> {
 
     List<InquiryAnswerAttachmentEntity> findAllByInquiryAnswerId(Long id);
