+++ src/main/java/life/mosu/mosuserver/presentation/inquiry/dto/InquiryAnswerUpdateRequest.java
@@ -1,7 +1,7 @@
 package life.mosu.mosuserver.presentation.inquiry.dto;
 
 import java.util.List;
-import life.mosu.mosuserver.presentation.faq.dto.FileRequest;
+import life.mosu.mosuserver.global.util.FileRequest;
 
 public record InquiryAnswerUpdateRequest(
         String title,
