+++ src/main/java/life/mosu/mosuserver/presentation/inquiry/dto/InquiryCreateRequest.java
@@ -3,7 +3,7 @@ package life.mosu.mosuserver.presentation.inquiry.dto;
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
 import life.mosu.mosuserver.domain.inquiry.InquiryJpaEntity;
-import life.mosu.mosuserver.presentation.faq.dto.FileRequest;
+import life.mosu.mosuserver.global.util.FileRequest;
 
 public record InquiryCreateRequest(
         @NotNull String title,
